adding script to explicitely set additionalProperties to False
By default, AdditionalProperties is set to `true` in the OpenAPI Spec. This is generating confusion on the API doc portal.

This script is automatically setting AdditionalProperties to `false` when it's not already defined (because already set to `false` or because it's actually used).
The script is excluding the "root" objects from the POST/PUT payload to allow the SDKs to accept AdditionalProperties, which are required to "remove" attributes from the root level